// src/client/api/generated/model/publishMode.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Spec Studio API
 * API local của Spec Studio. Lỗi chỉ trả mã (ErrorCode); FE tự dịch.
 * OpenAPI spec version: 1.0.0
 */

export type PublishMode = typeof PublishMode[keyof typeof PublishMode];


export const PublishMode = {
  PUSH: 'PUSH',
  PULL_REQUEST: 'PULL_REQUEST',
} as const;